+++ packages/shared/src/types/datetime.types.ts
@@ -1,5 +1,3 @@
-import {z} from 'zod';
-
 /**
  * A number representing a day of the week, where 0 is Sunday and 6 is the following Saturday.
  */
@@ -19,8 +17,3 @@ export interface TimeOfDay {
   readonly hour: number; // 0–23
   readonly minute: number; // 0–59
 }
-
-export const TimeOfDaySchema = z.object({
-  hour: z.number().int().min(0).max(23),
-  minute: z.number().int().min(0).max(59),
-});
