+++ packages/shared/src/types/userFeedSubscriptions.types.ts
@@ -2,7 +2,6 @@ import {z} from 'zod';
 
 import type {AccountId} from '@shared/types/accounts.types';
 import type {DeliverySchedule} from '@shared/types/deliverySchedules.types';
-import {DeliveryScheduleFromStorageSchema} from '@shared/types/deliverySchedules.types';
 import type {PersistedFeedSourceType} from '@shared/types/feedSourceTypes.types';
 import {FeedSourceType, PERSISTED_FEED_SOURCE_TYPES} from '@shared/types/feedSourceTypes.types';
 import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
@@ -10,6 +9,7 @@ import type {BaseStoreItem} from '@shared/types/utils.types';
 import {YouTubeChannelIdSchema, type YouTubeChannelId} from '@shared/types/youtube.types';
 
 import {AccountIdSchema} from '@shared/schemas/accounts.schema';
+import {DeliveryScheduleFromStorageSchema} from '@shared/schemas/deliverySchedules.schema';
 
 /**
  * Strongly-typed type for a {@link UserFeedSubscription}'s unique identifier. Prefer this over
